Move migrate package
